requirements/prod.txt
wheel==0.29.0
tweepy==3.5.0
Flask==0.12.2
python-dotenv==0.6.4
flask_oauthlib==0.9.4